Add files via upload
Overview
The Boolean satisfiability (SAT) problem involves determining whether a Boolean formula can be satisfied. This problem is a cornerstone in computer science as the classic NP-complete problem. Over the past decades, extensive research has been dedicated to solving SAT problems using both complete and incomplete solvers.

An advanced variant of this problem is Model Counting (#SAT), where the solver must compute the number of solutions for a given Boolean formula. Numerous solvers have been developed to address this challenge.

Project Goal
The aim of this project is to develop an Algorithm Selection (AS) approach for Model Counting. Given a #SAT instance, our goal is to identify the specific solver that performs best for that instance. Our machine learning approach will classify each instance to determine the most suitable solver.

Dataset
The dataset consists of a vector of 72 features for each #SAT problem instance, providing general information about the problem (e.g., number of variables, number of clauses). For each instance, there is also a 'label' column indicating the name of the optimal solver.